make ArtistTree instantiation less insane
previously, `ArtistTree::build` returned a `Graph` (instead of storing
it in the struct), which made it awkward for downstream methods
(primarily `as_dot`) to access. this was compounded by the fact that
`build` also required a `pool`. the end result was an awkward two-step
struct instantiation that i could never remember how to get right.

now, constructors of `ArtistTree` are only required to pass in a pool
once, and as early as possible (i.e. in `new`). the `Graph`, now stored
as a struct field, is instantiated and constructed in `new`.