replace petgraph::dot with graphviz_rust
petgraph::dot offers little support for customisation of the resulting
dot graph, which means we would have to modify the dot string directly
(or worse, the svg). this severely hampers styling and makes testing a
pain. graphviz_rust addresses this issue by providing an expressive API,
at the (small) expense of having to construct the dot by hand.