guard against empty `Vec`s when fetching from `artist_pairs`
because of the new charts `endpoint`, it is now possible for artists to
be added into `artists` table, without being added into the
`artist_pairs` table. thus, fetching rows from `artist_pairs` may result
in an empty `Vec`, leading to an empty graph. to prevent this, empty
`Vec`s are coerced into `None`, so they can be checked upstream.